src: change type size from int to size_t for Q_strncpyz & Q_strcat function

+++ src/cgame/cg_debriefing.c
@@ -2103,7 +2103,7 @@ void CG_Debriefing_ParseAwards(void)
 	char       *cs = (char *)CG_ConfigString(CS_ENDGAME_STATS);
 	const char *token;
 	char       *s;
-	int        size, len;
+	size_t     size, len;
 	int        clientNum;
 	float      value;
 	char       buffer[sizeof(cgs.dbAwardNamesBuffer)];
