+++ src/cgame/cg_main.c
@@ -2208,7 +2208,7 @@ const char *CG_ConfigString(int index)
  * @param[in] buffsize
  * @return
  */
-int CG_ConfigStringCopy(int index, char *buff, int buffsize)
+int CG_ConfigStringCopy(int index, char *buff, size_t buffsize)
 {
 	Q_strncpyz(buff, CG_ConfigString(index), buffsize);
 	return strlen(buff);
@@ -2971,9 +2971,9 @@ void QDECL CG_WriteToLog(const char *fmt, ...)
 	}
 	else
 	{
-		va_list      argptr;
-		char         string[1024];
-		unsigned int l;
+		va_list argptr;
+		char    string[1024];
+		size_t  l;
 
 		Com_sprintf(string, sizeof(string), "%s ", CG_GetRealTime());
