+++ src/game/g_misc.c
@@ -2703,7 +2703,7 @@ float AngleDifference(float ang1, float ang2)
  * @param[in] size
  * @return client name
  */
-char *ClientName(int client, char *name, int size)
+char *ClientName(int client, char *name, size_t size)
 {
 	char buf[MAX_INFO_STRING];
